Pull audio logic into service
